+++ lib/init.scm
@@ -11,10 +11,10 @@
                                     (if (null? e)
                                         (resolve (BrowserFS.BFSRequire "fs"))
                                         (reject e)))))))
-                 ((not (null? self.BrowserFS))
-                  (console.warn (string-append "BrowserFS is not initialized and "
-                                               "IndexedDB is not available"))
+                 ((not (indexed-db?))
+                  (console.warn (string-append "No FS found and IndexedDB "
+                                               "is not available"))
                   nil))))
   (let ((internal (lips.env.get '**internal-env**)))
     (if (not (null? fs))
-        (internal.set "fs" fs))))
+        (lips.set_fs fs))))