+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * add support for `(scheme-report-environment 7)`
 * add a way to compare custom instances with `equal?` [#270](https://github.com/jcubic/lips/issues/270)
 * add support for `equal?` on records [#273](https://github.com/jcubic/lips/issues/273)
+* expose `lips.set_fs` function [#256](https://github.com/jcubic/lips/issues/256)
 ### Bugfix
 * remove evaluating of async list data as first argument
 * fix `number->string` for binary numbers